Verify if TXT record is updated before returning.
There might be a slight delay between creating a record using the API
and the record actually being present in the NS of TransIP. This change
checks if the record is present before completing.

It verifies the root nameserver for the domain points to transip, and
while we're at it, now supports domains with multiple dots (e.g.
.co.uk).

It checks the root nameserver for the validation record, since that's
what letsencrypt does (and that bypasses caching issues). It tries every
5 seconds for up to a minute.

Lastly some formatting fixes.